pyproject.toml
[project]
name = "langchainProject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=0.0.6",
    "python-dotenv>=1.2.1",
]
